updating optimization on gradient descent and line search.
